Refactor website example includes
