+++ website/src/CMakeLists.txt
@@ -2,17 +2,18 @@
 # Copyright (c) Contributors to the OpenEXR Project
 
 add_executable(website_src
-    compression.cpp
     envmap.cpp
 
-    IStream.cpp
-    C_IStream.cpp
     C_IStream_read.cpp
     C_IStream_seekg.cpp
     C_IStream_tellg.cpp
     C_IStream_clear.cpp
+
+    all.cpp
 )
+
 target_link_libraries(website_src OpenEXR::OpenEXR)
+
 set_target_properties(website_src PROPERTIES
   RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/bin"
 )
