+++ website/src/C_IStream_seekg.cpp
@@ -5,7 +5,7 @@
 
 #include <cstdint>
 #include <cstdio>
-#include "C_IStream.cpp"
+#include "C_IStream.h"
 
 void
 C_IStream::seekg (uint64_t pos)
